feat: separate EventHandler module
