+++ packages/core/src/types.ts
@@ -1,0 +1,6 @@
+export interface DrawingEventHandler {
+  start: () => void
+  end: () => void
+  move: (x: number, y: number) => void
+  resize?: (x: number, y: number) => void
+}
